pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "galleon"
version = "0.1.0"
description = "Electronic component sourcing agent"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
]

[project.scripts]
galleon = "galleon.cli:app"

[tool.setuptools.packages.find]
where = ["."]
